+++ src/Publish/Migrations/2014_06_29_122019_acl_users.php
@@ -8,6 +8,17 @@ class AclUsers extends Migration
 {
 
     /**
+     * Detect PostgreSQL database
+     *
+     * @return bool
+     */
+    public function isPGSQL()
+    {
+        $driver = \Config::get('database.default');
+        return \Config::get("database.connections.{$driver}.driver") === 'pgsql';
+    }
+
+    /**
      * Run the migrations.
      *
      * @return void
@@ -17,11 +28,17 @@ class AclUsers extends Migration
         if (!Schema::hasTable('acl_users')) {
             Schema::create('acl_users', function ($table) {
                 $table->increments('id');
-                $table->timestamps();
                 $table->string('login', 255);
                 $table->string('password', 255);
                 $table->string('group_id', 255);
-                $table->string('remember_token', 150);
+                $table->string('remember_token', 150)->nullable();
+
+                if ($this->isPGSQL()) {
+                    $table->timestamp('created_at')->default(DB::raw('now()::timestamp(0)'));
+                    $table->timestamp('updated_at')->default(DB::raw('now()::timestamp(0)'));
+                } else {
+                    $table->timestamps();
+                }
             });
 
             /**