+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 #v2.1.0
 
  * [[#7]](https://github.com/signes-pl/laravel-acl/issues/7) Updated Illuminate dependencies. Used caret operator.
+ * [[#6]](https://github.com/signes-pl/laravel-acl/issues/6) Changed mechanism to create timestamp fields for PostgreSQL database driver